+++ airfoileditor/__init__.py
@@ -1,3 +1,2 @@
 
 # package metadata
-__author__       = "Jochen Guenzel"                     # hatch "author"